Make log messages more user-friendly
Omit class names in the stdout log (in the Agent.log file they remain intact). Fix the agent to wait for the installation of the on-device components on all connected devices to finish and then show a "has started successfully" message which should be easier to spot by the user.