Fix NullPointerException caused by the FileRecycler
Make FileRecycler static, so it is shared between DeviceManager instances and does not cause NullPointerException when a device is plugged in while the Agent is running. Fix a bug in the FileRecycler which was making it clear files indefinitely. Also update the isoparser library version to potentially fix a memory leak inside the library.